Extract Header component and relocate shared UI to app/ui, closes #23
