+++ TASKS.md
@@ -22,5 +22,5 @@
 - [x] add /learn route and dashboard ui rendering first teaching module, #19
 - [x] define module schema and author module content for simple spa app, #20
 - [x] wire selection to load module files into the editor and file tree, #21
-- [ ] extract dashboard, project, and page headers as standalone components, #23
+- [x] extract dashboard, project, page headers as standalone components, #23
 - [ ] add App Router navigation between editor and dashboard interfaces, #22
