+++ src/components/FileTree.tsx
@@ -1,4 +1,4 @@
-import FileIcon from '@/components/FileIcon'
+import FileIcon from '@/app/ui/FileIcon'
 import type { File } from '@/types/index'
 import { FileTreeProps } from '@/types/components'
